[without_modulo] store length, not write index
This results in a ~35% performance increase across the board, making it
competitive with `AllocRingBuffer` and beating `AllocRingBuffer` for
non-power-of-two sizes.  Apparently, tracking the base and length
instead of read and write pointers requires less logic and is optimized
better - probably because the base and length change less often (and in
less complex ways).